+++ knowthenics-web/src/main/java/ru/dpopkov/knowthenics/controllers/AnswerController.java
@@ -1,14 +1,23 @@
 package ru.dpopkov.knowthenics.controllers;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
+import ru.dpopkov.knowthenics.services.AnswerService;
 
 @Controller
 @RequestMapping("/answers")
 public class AnswerController {
 
+    private final AnswerService answerService;
+
+    public AnswerController(AnswerService answerService) {
+        this.answerService = answerService;
+    }
+
     @RequestMapping({"", "/", "list", "list.html", "index", "index.html"})
-    public String list() {
+    public String list(Model model) {
+        model.addAttribute("answers", answerService.findAll());
         return "answers/list";
     }
 }