+++ knowthenics-web/src/main/java/ru/dpopkov/knowthenics/controllers/CategoryController.java
@@ -1,14 +1,23 @@
 package ru.dpopkov.knowthenics.controllers;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
+import ru.dpopkov.knowthenics.services.CategoryService;
 
 @Controller
 @RequestMapping("/categories")
 public class CategoryController {
 
+    private final CategoryService categoryService;
+
+    public CategoryController(CategoryService categoryService) {
+        this.categoryService = categoryService;
+    }
+
     @RequestMapping({"", "/", "list", "list.html", "index", "index.html"})
-    public String list() {
+    public String list(Model model) {
+        model.addAttribute("categories", categoryService.findAll());
         return "categories/list";
     }
 }