+++ knowthenics-web/src/main/java/ru/dpopkov/knowthenics/controllers/QuestionsController.java
@@ -1,14 +1,23 @@
 package ru.dpopkov.knowthenics.controllers;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
+import ru.dpopkov.knowthenics.services.QuestionService;
 
 @Controller
 @RequestMapping("/questions")
 public class QuestionsController {
 
+    private final QuestionService questionService;
+
+    public QuestionsController(QuestionService questionService) {
+        this.questionService = questionService;
+    }
+
     @RequestMapping({"", "/", "list", "list.html", "index", "index.html"})
-    public String list() {
+    public String list(Model model) {
+        model.addAttribute("questions", questionService.findAll());
         return "questions/list";
     }
 }