+++ knowthenics-web/src/main/java/ru/dpopkov/knowthenics/controllers/SourcesController.java
@@ -1,14 +1,23 @@
 package ru.dpopkov.knowthenics.controllers;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
+import ru.dpopkov.knowthenics.services.SourceService;
 
 @Controller
 @RequestMapping("/sources")
 public class SourcesController {
 
+    private final SourceService sourceService;
+
+    public SourcesController(SourceService sourceService) {
+        this.sourceService = sourceService;
+    }
+
     @RequestMapping({"", "/", "list", "list.html", "index", "index.html"})
-    public String list() {
+    public String list(Model model) {
+        model.addAttribute("sources", sourceService.findAll());
         return "sources/list";
     }
 }